Make debug system command work on all platforms
Move containerd-specific code to debug_system_linux.go and provide
a stub for non-Linux platforms in debug_system_other.go.
