+++ cli/commands/debug_system_other.go
@@ -1,0 +1,7 @@
+//go:build !linux
+
+package commands
+
+func gatherContainers(ctx *Context, socket, namespace string) ([]byte, error) {
+	return []byte("containerd container listing is only supported on Linux\n"), nil
+}
